+++ Classes/Configuration/FlexForms.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace OliverKlee\Onetimeaccount\Configuration;
 
+use TYPO3\CMS\Core\Information\Typo3Version;
+
 /**
  * This class provides functions for building FlexForms.
  */
@@ -38,15 +40,25 @@ class FlexForms
     /**
      * Sets the selectable items for the fields to display in `$configuration`.
      *
-     * @param array<string, array<string, string>> $configuration
+     * @param array<string, mixed> $configuration
      */
     public function buildFields(array &$configuration): void
     {
-        /** @var array<int, array{0: non-empty-string, 1: non-empty-string}> $items */
+        if ((new Typo3Version())->getMajorVersion() < 12) {
+            $labelKey = 0;
+            $valueKey = 1;
+        } else {
+            $labelKey = 'label';
+            $valueKey = 'value';
+        }
+
         $items = [];
         foreach (static::AVAILABLE_FIELDS as $fieldKey) {
             $label = 'LLL:EXT:onetimeaccount/Resources/Private/Language/locallang.xlf:' . $fieldKey;
-            $items[] = [$label, $fieldKey];
+            $items[] = [
+                $labelKey => $label,
+                $valueKey => $fieldKey,
+            ];
         }
 
         $configuration['items'] = $items;